All i18n functions moved back to single file to work in the same scope
